refactor to ScenarioTransformation
